backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.2
pydantic-settings==2.6.1
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.35.68
pypdf==5.1.0
python-docx==1.1.2
langchain==0.3.8
langchain-openai==0.2.9
langchain-chroma==0.1.4
chromadb==0.5.20
openai==1.55.3
slowapi==0.1.9
reportlab==4.2.5
email-validator==2.2.0
